+++ sanyan/__init__.py
@@ -13,4 +13,4 @@
 
 from __future__ import annotations
 
-__version__ = '3.56.1'
+__version__ = '3.56.2'